chore: move OwnedObjectPreDeleteHook to pkg/utils/kubernetes/reduce
The OwnedObjectPreDeleteHook helper lives in controller/pkg/dataplane,
but its only consumers are the generic resource-reduction machinery and
controller/pkg/secrets. pkg/utils/kubernetes/reduce already defines the
PreDeleteHook type it implements, so the helper belongs there.

Moving it also keeps the upcoming shared specialized-DataPlane
reconciler package (controller/pkg/dataplane) free of a dependency in
the opposite direction: controller/pkg/secrets must not import the
reconciler package, which would otherwise create an import cycle in
tests once the reconciler package grows.

This is a pure move plus import updates; no behavior change.

Part of a PR stack splitting the AIGateway/EventGateway DataPlane
reconciler deduplication (#4842) into reviewable layers.
