+++ controller/dataplane/owned_resources_test.go
@@ -15,10 +15,10 @@ import (
 
 	operatorv1beta1 "github.com/kong/kong-operator/v2/api/gateway-operator/v1beta1"
 	"github.com/kong/kong-operator/v2/controller/pkg/builder"
-	"github.com/kong/kong-operator/v2/controller/pkg/dataplane"
 	"github.com/kong/kong-operator/v2/controller/pkg/op"
 	"github.com/kong/kong-operator/v2/pkg/consts"
 	k8sutils "github.com/kong/kong-operator/v2/pkg/utils/kubernetes"
+	k8sreduce "github.com/kong/kong-operator/v2/pkg/utils/kubernetes/reduce"
 	k8sresources "github.com/kong/kong-operator/v2/pkg/utils/kubernetes/resources"
 )
 
@@ -46,7 +46,7 @@ func TestEnsureIngressServiceForDataPlane(t *testing.T) {
 				WithIngressServiceType(corev1.ServiceTypeLoadBalancer).
 				Build(),
 			existingServiceModifier: func(t *testing.T, ctx context.Context, c client.Client, svc *corev1.Service) {
-				require.NoError(t, dataplane.OwnedObjectPreDeleteHook(ctx, c, svc))
+				require.NoError(t, k8sreduce.OwnedObjectPreDeleteHook(ctx, c, svc))
 				require.NoError(t, c.Delete(ctx, svc))
 			},
 			expectedCreatedOrUpdated: op.Created,
@@ -277,7 +277,7 @@ func TestEnsureIngressServiceForDataPlane(t *testing.T) {
 				WithIngressServiceType(corev1.ServiceTypeLoadBalancer).
 				Build(),
 			existingServiceModifier: func(t *testing.T, ctx context.Context, c client.Client, svc *corev1.Service) {
-				require.NoError(t, dataplane.OwnedObjectPreDeleteHook(ctx, c, svc))
+				require.NoError(t, k8sreduce.OwnedObjectPreDeleteHook(ctx, c, svc))
 				require.NoError(t, c.Delete(ctx, svc))
 			},
 			expectedCreatedOrUpdated: op.Created,
