+++ pkg/utils/kubernetes/reduce/pre_delete_hook.go
@@ -1,4 +1,4 @@
-package dataplane
+package reduce
 
 import (
 	"context"
